requirements.txt
astroid==2.13.2
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
dill==0.3.6
distro==1.8.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.4
isort==5.11.4
jsonschema==3.2.0
lazy-object-proxy==1.9.0
lxml==4.9.2
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==23.0
paramiko==2.12.0
pathspec==0.10.3
platformdirs==2.6.2
psycopg2-binary==2.9.5
pycparser==2.21
pylint==2.15.10
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dotenv==0.21.0
PyYAML==5.4.1
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
texttable==1.6.7
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.14
websocket-client==0.59.0
wrapt==1.14.1
